finally, self-installable!

+++ out/repo_files/busybox/post.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
-rm $1/bin/ln
-busybox ln -s busybox $1/bin/ln
-rm $1/bin/old_busybox
+if [ -f $1/bin/old_busybox ]; then
+ rm $1/bin/ln
+ busybox ln -s busybox $1/bin/ln
+ rm $1/bin/old_busybox
+fi
